Refactor CollabTextArea component location and update import path
